+++ lib/auth0/auth0.dart
@@ -1,3 +1,4 @@
 export 'get_logged_in_user/get_logged_in_user.dart';
+export 'handle_redirect_callback/handle_redirect_callback.dart';
 export 'login_user/login_user.dart';
 export 'logout_user/logout_user.dart';