+++ lib/auth0/get_logged_in_user/get_logged_in_user.dart
@@ -6,8 +6,10 @@ import 'get_logged_in_user_stub.dart'
 Future<String?> getLoggedInUserId({
   required String auth0Domain,
   required String auth0ClientId,
+  String? redirectUri,
 }) =>
     impl.getLoggedInUserId(
       auth0Domain: auth0Domain,
       auth0ClientId: auth0ClientId,
+      redirectUri: redirectUri,
     );