+++ lib/auth0/get_logged_in_user/get_logged_in_user_mobile.dart
@@ -5,9 +5,14 @@ import 'package:auth0_flutter2/auth0/init_auth0/init_auth0_mobile.dart';
 Future<String?> getLoggedInUserId({
   required String auth0Domain,
   required String auth0ClientId,
+  String? redirectUri,
 }) async {
-  Auth0 auth0 = initAuth0Mobile(auth0Domain, auth0ClientId);
   try {
+    Auth0 auth0 = initAuth0Mobile(
+      auth0Domain,
+      auth0ClientId,
+    );
+
     /// Check if user is logged in.
     if (await auth0.credentialsManager.hasValidCredentials()) {
       var creds = await auth0.credentialsManager.credentials();