+++ lib/auth0/get_logged_in_user/get_logged_in_user_web.dart
@@ -6,13 +6,15 @@ import 'package:auth0_flutter2/auth0/auth0_flutter_web/auth0_flutter_web.dart'
 Future<String?> getLoggedInUserId({
   required String auth0Domain,
   required String auth0ClientId,
+  String? redirectUri,
 }) async {
-  auth0_web.Auth0 auth0 = await initAuth0Web(
-    auth0Domain,
-    auth0ClientId,
-  );
-
   try {
+    auth0_web.Auth0 auth0 = await initAuth0Web(
+      auth0Domain: auth0Domain,
+      auth0ClientId: auth0ClientId,
+      redirectUri: redirectUri,
+    );
+
     var credentials = await auth0.getUser();
     return credentials?["sub"];
   } catch (e) {