+++ lib/auth0/handle_redirect_callback/handle_redirect_callback_mobile.dart
@@ -1,0 +1,9 @@
+/// Not applicable for mobile.
+Future<bool> handleRedirectCallback({
+  required String auth0Domain,
+  required String auth0ClientId,
+  String? redirectUri,
+  String? url,
+}) async {
+  return true;
+}