+++ lib/auth0/handle_redirect_callback/handle_redirect_callback_stub.dart
@@ -1,8 +1,9 @@
-/// Stub method for getLoggedInUserId.
-Future<String?> getLoggedInUserId({
+/// Stub method for logging in a user.
+Future<bool> handleRedirectCallback({
   required String auth0Domain,
   required String auth0ClientId,
   String? redirectUri,
+  String? url,
 }) =>
     throw UnsupportedError(
       'This is the stub method, use the active method instead.',