+++ lib/auth0/login_user/login_user_web.dart
@@ -1,5 +1,4 @@
 import 'package:auth0_flutter2/auth0/auth0.dart';
-import 'package:auth0_flutter2/auth0/auth0_flutter_web/src/options.dart';
 import 'package:auth0_flutter2/auth0/init_auth0/init_auth0_web.dart';
 import 'package:auth0_flutter2/auth0/auth0_flutter_web/auth0_flutter_web.dart'
     as auth0_web;
@@ -8,11 +7,13 @@ import 'package:auth0_flutter2/auth0/auth0_flutter_web/auth0_flutter_web.dart'
 Future<String?> loginUser({
   required String auth0Domain,
   required String auth0ClientId,
+  required String? redirectUri,
   Future<void> Function()? afterLogin,
 }) async {
   auth0_web.Auth0? auth0 = await initAuth0Web(
-    auth0Domain,
-    auth0ClientId,
+    auth0Domain: auth0Domain,
+    auth0ClientId: auth0ClientId,
+    redirectUri: redirectUri,
   );
 
   try {
@@ -24,9 +25,7 @@ Future<String?> loginUser({
 
     // If user is not logged in, direct to auth page.
     if (user == null) {
-      await auth0.loginWithPopup(
-        options: PopupLoginOptions(),
-      );
+      await auth0.loginWithRedirect(redirectUri: redirectUri ?? "");
 
       // Check again if user is logged in (after auth attempt).
       final nowUser = await getLoggedInUserId(
@@ -35,7 +34,6 @@ Future<String?> loginUser({
       );
 
       if (nowUser != null) {
-        if (afterLogin != null) afterLogin();
         return nowUser;
       } else {
         return null;