+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.0.6] - 11.12.2022
+
+* Refactor web login to use a redirect instead of a popup. Add additional property to Auth0Flutter2 class, redirectUri, to support new web login functionality.
+
+## [0.0.5] - 06.12.2022
+
+* Update package description.
+
 ## [0.0.4] - 06.12.2022
 
 * So close to the 140/140 Pub Points we can taste it! More convertion from comments to dartdoc comments.